requirements.txt
flask
google-genai>=1.10.0
python-dotenv
gunicorn
